+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `EntropyBlockingRandomWrapper`: won't return any output when below a minimum entropy (recommended to use 0 or less).
+  Can be configured with a `SeedGenerator` to reseed on the same thread in the absence of a working
+  `RandomSeederThread`.
+
 ### Changed
 
 - slf4j loggers are now initialized lazily, since they're often unused. This should speed up startup.